Add files via upload
The code will select three area from a galaxy, based on the input (the position of the region: center, intermediate region or, outskirt) from the user. Based on that, it can chose an area size, that will also be user dependent (it will ask for the amount of area, the user want to work on). Then selecting the region, the code, will return properties for each of them. The plots can be plotted using them.