Add files via upload
Using this code, the  user can give the co-ordinates as input, and also, the amount of area around that point, on which the user want to work on, can be given as input by the user. The "mask" can be used as indices to get the properties for the selected regions.